+++ runtime/server/queries_test.go
@@ -136,7 +136,7 @@ func TestServer_UpdateLimit_UNION(t *testing.T) {
 func prepareOLAPStore(t *testing.T) drivers.OLAPStore {
 	conn, err := duckdb.Driver{}.Open(map[string]any{"dsn": "?access_mode=read_write&rill_pool_size=4"}, zap.NewNop())
 	require.NoError(t, err)
-	olap, ok := conn.AsOLAPStore()
+	olap, ok := conn.AsOLAP()
 	require.True(t, ok)
 	return olap
 }
