+++ runtime/testruntime/testruntime.go
@@ -16,7 +16,6 @@ import (
 	// Load database drivers for testing.
 	_ "github.com/rilldata/rill/runtime/drivers/duckdb"
 	_ "github.com/rilldata/rill/runtime/drivers/file"
-	_ "github.com/rilldata/rill/runtime/drivers/localfile"
 	_ "github.com/rilldata/rill/runtime/drivers/sqlite"
 )
